chore: setup typescript, linting, formatting and testing
Setup the project's tooling and scripts.

Linting and formatting is handled with ESLint and Prettier with custom
configs which only slightly extend conventional defaults. Husky ensures
code is linted and formatted before committed.

Testing is done with Jest and Testing Library.

At this point, the project is ready to go for test driven development
and continuous integration deployments.